… src/Servers/Kestrel/Core/src/KestrelServerOptions.cs
… src/Servers/Kestrel/Core/test/HttpRequestHeadersTests.cs
… src/Servers/Kestrel/Core/test/KestrelServerOptionsTests.cs
… src/Servers/Kestrel/Core/test/KestrelServerTests.cs
… src/Servers/Kestrel/Core/test/UTF8Decoding.cs
… src/Servers/Kestrel/Kestrel/test/KestrelConfigurationLoaderTests.cs
… src/Servers/Kestrel/perf/Kestrel.Performance/BytesToStringBenchmark.cs
… src/Servers/Kestrel/shared/KnownHeaders.cs
… src/Servers/Kestrel/test/InMemory.FunctionalTests/Http2/Http2StreamTests.cs
… src/Servers/Kestrel/test/InMemory.FunctionalTests/Http2/Http2TestBase.cs
… src/Servers/Kestrel/test/InMemory.FunctionalTests/Http3/Http3TestBase.cs
… src/Servers/Kestrel/test/InMemory.FunctionalTests/RequestTests.cs
… src/Shared/ServerInfrastructure/StringUtilities.cs