- TODONE for this PR:
  - Benchmarks. I want to make sure we agree on this design before benchmarking. In the fast path, it only adds a couple comparisons, so I think we should be good.

---

- TODO for future PRs:
  - URL decoding
  - ANCM in-proc support


Partially Addresses #17400