@halter73 2020-06-26
# Http1ConnectionBenchmark Microbenchmark Results

## Before (release/5.0-preview7)

#### default (UTF-8) encoding

|     Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| LiveAspNet | 1.119 us | 0.0222 us | 0.0458 us | 894,046.9 | 0.0210 |     - |     - |     880 B |

#### latin1 (ISO-8859-1) encoding

|     Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| LiveAspNet | 1.171 us | 0.0231 us | 0.0549 us | 854,282.0 | 0.0210 |     - |     - |     880 B |

## After (halter73/17400)

#### default (UTF-8) encoding
|     Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| LiveAspNet | 1.135 us | 0.0180 us | 0.0160 us | 881,286.2 | 0.0229 |     - |     - |     880 B |

#### latin1 (ISO-8859-1) encoding (`RequestHeaderEncodingSelector = _ => Encoding.Latin1`)
|     Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| LiveAspNet | 1.172 us | 0.0217 us | 0.0203 us | 853,229.7 | 0.0210 |     - |     - |     880 B |

#### Alternative default (UTF-8) encoding (`RequestHeaderEncodingSelector = _ => null`)
|     Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| LiveAspNet | 1.146 us | 0.0166 us | 0.0155 us | 872,465.7 | 0.0229 |     - |     - |     880 B |

#### Alternative defaultish (UTF-8) encoding (`RequestHeaderEncodingSelector = _ => Encoding.UTF8`)
|     Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| LiveAspNet | 1.310 us | 0.0173 us | 0.0153 us | 763,159.7 | 0.0229 |     - |     - |     880 B |

---

Given that we're trying to not regress perf rather than improve it with this PR, this all looks reasonable to me.

I made sure to set `DisableStringReuse = true` and hard-coded the "runtimeFrameworkVersion" to 5.0.0-preview.8.20323.9 for all benchmark runs. See https://github.com/dotnet/aspnetcore/pull/23000#discussion_r446348499