@davidfowl 2020-06-26
Can you run a CPU profile to see if there's anything that stands out?