@davidfowl 2020-06-26
Also can you run the PR benchmark? This is basically in the hot path.