@halter73 2020-06-26
@aspnet-hello benchmark

> Can you run a CPU profile to see if there's anything that stands out?

I don't have time for this today, and I still want this in preview7. It wouldn't be that useful yet anyway, since in the default case, we do 1 extra ReferenceEquals() comparison per request header and follow the old request decoding path. I don't need a CPU profile to tell me that.