@halter73 2020-06-27
I'm also talking to @sebastienros about the benchmark server issue. I agree with him that we should move over to the new driver so we can use the same servers as everything else. We just need to budget time to make the changes.